+++ sidebar/src/components/UI/molecules/bucket_create_button/style.js
@@ -1,0 +1,9 @@
+const style = {
+  minWidth: '100%',
+  minHeight: '40px',
+  backgroundColor: '#454552',
+  color: 'white',
+  fontSize: 20,
+};
+
+export default style;
